+++ patterns.test.js
@@ -1,18 +1,18 @@
 import test from 'ava';
-import {patternToRegex, globToRegex} from './index.js';
+import {patternToRegex} from './index.js';
 
-function macroPattern(t, pattern, matching) {
+function macro(t, pattern, matching) {
 	const regex = patternToRegex(pattern);
 	for (const url of matching) {
 		t.regex(url, regex);
 	}
 }
 
-macroPattern.title = (_, pattern) => pattern;
+macro.title = (_, pattern) => pattern;
 
 // Patterns pulled from https://developer.chrome.com/extensions/match_patterns
-const patternMap = new Map();
-patternMap.set('http://*/*', [
+const map = new Map();
+map.set('http://*/*', [
 	'http://www.google.com/',
 	'http://example.org/foo/bar.html',
 	'http://fregante.com',
@@ -20,34 +20,34 @@ patternMap.set('http://*/*', [
 	'http://www.fregante.com/',
 	'http://www.fregante.com/mail',
 ]);
-patternMap.set('http://*/foo*', [
+map.set('http://*/foo*', [
 	'http://example.com/foo/bar.html',
 	'http://www.google.com/foo',
 	'http://fregante.com/foobar',
 	'http://mail.fregante.com/foo/king',
 ]);
-patternMap.set('https://*.google.com/foo*bar', [
+map.set('https://*.google.com/foo*bar', [
 	'https://www.google.com/foo/baz/bar',
 	'https://docs.google.com/foobar',
 	'https://google.com/foonderbar',
 ]);
-patternMap.set('http://example.org/foo/bar.html', [
+map.set('http://example.org/foo/bar.html', [
 	'http://example.org/foo/bar.html',
 ]);
-patternMap.set('file:///foo*', [
+map.set('file:///foo*', [
 	'file:///foo/bar.html',
 	'file:///foo',
 ]);
-patternMap.set('http://127.0.0.1/*', [
+map.set('http://127.0.0.1/*', [
 	'http://127.0.0.1/',
 	'http://127.0.0.1/foo/bar.html ',
 ]);
-patternMap.set('*://mail.google.com/*', [
+map.set('*://mail.google.com/*', [
 	'http://mail.google.com/foo/baz/bar',
 	'https://mail.google.com/foobar',
 ]);
 
-patternMap.set('<all_urls>', [
+map.set('<all_urls>', [
 	'http://mail.google.com/foo/baz/bar',
 	'https://mail.google.com/foobar',
 	'file:///foo/bar.html',
@@ -56,8 +56,8 @@ patternMap.set('<all_urls>', [
 	'ftp://example.com',
 ]);
 
-for (const [pattern, urls] of patternMap) {
-	test(macroPattern, pattern, urls);
+for (const [pattern, urls] of map) {
+	test(macro, pattern, urls);
 }
 
 test('Should not match anything if no patterns are passed', t => {
@@ -90,33 +90,4 @@ for (const pattern of invalidPatternsThatPass) {
 			message: /is an invalid pattern, it must match/,
 		});
 	});
-}
-
-function macroGlob(t, glob, matching) {
-	const regex = globToRegex(glob);
-	for (const url of matching) {
-		t.regex(url, regex);
-	}
-}
-
-macroGlob.title = (_, glob) => glob;
-
-// Pulled from https://developer.chrome.com/docs/extensions/mv3/content_scripts/#matchAndGlob
-const globMap = new Map();
-globMap.set('https://???.example.com/foo/*', [
-	'https://www.example.com/foo/bar',
-	'https://the.example.com/foo/',
-]);
-
-globMap.set('*nytimes.com/???s/*', [
-	'https://www.nytimes.com/arts/index.html',
-	'https://www.nytimes.com/jobs/index.html',
-]);
-
-globMap.set('*google.com*', [
-	'https://google.com/',
-]);
-
-for (const [pattern, urls] of globMap) {
-	test(macroGlob, pattern, urls);
 }
